LayoutTests/platform/glib/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-request-redirect.https-expected.txt

CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: XMLHttpRequest cannot load https://web-platform.test:9443/service-workers/service-worker/resources/?url=%2Fservice-workers%2Fservice-worker%2Fresources%2Fredirect.py%3FRedirect%3D%252Fservice-workers%252Fservice-worker%252Fresources%252Fsimple.txt&redirect-mode=error.
CONSOLE MESSAGE: Response served by service worker is opaque redirect
CONSOLE MESSAGE: XMLHttpRequest cannot load https://web-platform.test:9443/service-workers/service-worker/resources/redirect.py?Redirect=%2Fservice-workers%2Fservice-worker%2Fresources%2Fsimple.txt due to access control checks.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/?url=%2Fservice-workers%2Fservice-worker%2Fresources%2Fredirect.py%3FRedirect%3D%252Fservice-workers%252Fservice-worker%252Fresources%252Fsquare.png&redirect-mode=error.
CONSOLE MESSAGE: Response served by service worker is opaque redirect
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/redirect.py?Redirect=%2Fservice-workers%2Fservice-worker%2Fresources%2Fsquare.png.
CONSOLE MESSAGE: Response served by service worker is opaque redirect
CONSOLE MESSAGE: Cannot load image https://web-platform.test:9443/service-workers/service-worker/resources/?url=%2Fservice-workers%2Fservice-worker%2Fresources%2Fredirect.py%3FRedirect%3D%252Fservice-workers%252Fservice-worker%252Fresources%252Fsquare.png&redirect-mode=manual due to access control checks.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/?url=%2Fservice-workers%2Fservice-worker%2Fresources%2Fredirect.py%3FRedirect%3D%252Fmedia%252Fsound_5.oga&redirect-mode=error.
CONSOLE MESSAGE: Response served by service worker is opaque redirect
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/redirect.py?Redirect=%2Fmedia%2Fsound_5.oga.
CONSOLE MESSAGE: Response served by service worker is opaque redirect
CONSOLE MESSAGE: Response served by service worker has redirections
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample.html.
CONSOLE MESSAGE: Response served by service worker has redirections
CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/fetch-request-redirect-iframe.html?url=%2Fservice-workers%2Fservice-worker%2Fresources%2Fredirect.py%3FRedirect%3D%252Fservice-workers%252Fservice-worker%252Fresources%252Fsample.html&redirect-mode=error.
CONSOLE MESSAGE: Response served by service worker is opaque redirect
CONSOLE MESSAGE: XMLHttpRequest cannot load https://web-platform.test:9443/service-workers/service-worker/resources/redirect.py?Redirect=%2Fservice-workers%2Fservice-worker%2Fresources%2Fsimple.txt due to access control checks.
CONSOLE MESSAGE: Response served by service worker is opaque redirect
CONSOLE MESSAGE: XMLHttpRequest cannot load https://web-platform.test:9443/service-workers/service-worker/resources/redirect.py?Redirect=%2Fservice-workers%2Fservice-worker%2Fresources%2Fsimple.txt due to access control checks.

PASS Verify redirect mode of Fetch API and ServiceWorker FetchEvent.
PASS Verify redirected of Response(Fetch API) and ServiceWorker FetchEvent.
PASS Verify redirected of Response(Fetch API), Cache API and ServiceWorker FetchEvent.
PASS Normal XHR should succeed. - default
PASS Redirected XHR should succeed. - default
PASS Redirected XHR with Request.redirect=follow should succeed. - default
PASS Redirected XHR with Request.redirect=error should fail. - default
PASS Redirected XHR with Request.redirect=manual should fail. - default
PASS Normal image resource should be loaded. - default
PASS Redirected image resource should be loaded. - default
PASS Loading redirected image with Request.redirect=follow should succeed. - default
PASS Loading redirected image with Request.redirect=error should fail. - default
PASS Loading redirected image with Request.redirect=manual should fail. - default
PASS Normal audio resource should be loaded. - default
PASS Redirected audio resource should be loaded. - default
PASS Loading redirected audio with Request.redirect=follow should succeed. - default
PASS Loading redirected audio with Request.redirect=error should fail. - default
PASS Loading redirected audio with Request.redirect=manual should fail. - default
PASS Normal iframe loading should succeed. - default
PASS Normal redirected iframe loading should succeed. - default
PASS Redirected iframe loading with Request.redirect=follow should fail. - default
PASS Redirected iframe loading with Request.redirect=error should fail. - default
PASS Redirected iframe loading with Request.redirect=manual should succeed. - default
PASS Normal XHR should be resolved and response should not be redirected. - redirected
PASS Redirected XHR should be resolved and response should be redirected. - redirected
PASS Normal XHR should be resolved and response should not be redirected even with CORS mode. - redirected
PASS Redirected XHR should be resolved and response.redirected should be redirected with CORS mode. - redirected
PASS Redirected XHR should be reject and response should be redirected with NO-CORS mode and redirect-mode=manual. - redirected
PASS Redirected CORS image should be reject and response should not be redirected with NO-CORS mode. - redirected
PASS Normal XHR should be resolved and response should not be redirected. - cache
PASS Redirected XHR should be resolved and response should be redirected. - cache
PASS Normal XHR should be resolved and response should not be redirected even with CORS mode. - cache
PASS Redirected XHR should be resolved and response.redirected should be redirected with CORS mode. - cache
PASS Redirected XHR should be reject and response should be redirected with NO-CORS mode and redirect-mode=manual. - cache
PASS Redirected CORS image should be reject and response should not be redirected with NO-CORS mode. - cache
